+++ NEWS.md
@@ -3,6 +3,7 @@
 ## Development
 
 * `coef_rename()` gets a `poly=TRUE` argument to rename `poly(x, 2)`-style coefficients. Thanks to @mccarthy-m-g for code contribution #778.
+* `get_gof()`: `logLik` column converted to numeric for consistent types. Issue 649 reported on the `mice` Github.
 
 ## 2.1.1